Change the sampling algorithm for --auxinput files.
We always guaranteed that a single fcl file will use distinct
--auxinput files.  Now we also guarantee that the files are distinct
across the fcl dataset, provided there are enough inputs.  If we need
to re-sample the list of input files, we try to use each of the inputs
the same number of times, instead of allowing for the long Poisson
tails.