Do not require mu2ebintools for future versions of FHICLCPP.
mu2eprodsys migrated from fhicl-getpar (from mu2ebintools) to
fhicl-get (from fhiclcpp), and we do not need bintools for running
with current and future Offline (modern FHICLCPP) versions.  When
running an old version of Offline under mu2eprodsys v7_00_00 or later
we need to provide (or shadow) the fhicl-get command; export
a shell function for that as a part of UPS setup.